pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fwork-learnability"
version = "0.1.0"
description = "LLM-assisted framework learnability benchmark"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Research Team"}
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "lifelines>=0.28.0",  # Survival analysis
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "tiktoken>=0.5.0",  # Token counting
    "beautifulsoup4>=4.12.0",  # Doc scraping
    "markdownify>=0.11.0",  # HTML to markdown
    "rich>=13.0.0",  # CLI output
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
# Framework-specific dependencies (installed in separate venvs)
pydanticai = ["pydantic-ai>=1.0.0"]
haystack = ["haystack-ai>=2.0.0"]
langgraph = ["langgraph>=1.0.0"]
openai-agents = ["openai-agents>=0.1.0"]

[project.scripts]
fwork-learn = "scripts.run_experiment:main"
fetch-docs = "scripts.fetch_docs:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
